requirements.txt
pygame-ce~=2.3.0
pymunk~=6.5.1
numpy~=1.25.1
PyYAML~=6.0.1
